requirements.txt
aiohttp
pandas
polars
numpy
dash
plotly
python-binance
python-telegram-bot
apscheduler
pydantic
